DofStorageInterface added. This Class is the Interface of what we can expect to be a Dofstorage, i.e. provides a local dof iterator and a global dof iterator.
DofStorageDefault provides two extra featrures with default implementation.
All Dofstorages must be derived from this class but implementing only the
features of the above class.

The pwlinfunc.hh provides the implementation of linear base function with
polynom oeder 1 for triangles and a LagrangeMapper. Furthermore a
implementation of the corresponding function space.

For detailed description see the .hh files.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@102 419231ca-4411-0410-aba2-fc6cf8c4f609