Quadrature now with polynomial order as template parameter and as dynamic value. FastQuadrature is the template implementation and Quadrature the dynamic version.
gaussquadimp.hh .cc  is the GaussQuadrature implemented by Peter
but now with order as template parameter but the same functionality.
This was needed for the new implementation.

quadlqh.hh provides gauss quadrature with the new interface for lines,
quadrilaterlas and hexahedrons, i.e. its an interface between Peters
GaussQuadrature and the other stuff.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@107 419231ca-4411-0410-aba2-fc6cf8c4f609