AlbertGrid:
Jacobian_inverse implementation revised and now nearly as fast as the
Albert implementation.

NeighborIterator revised and now we have a speedup.

LagrangeBase:

caching of basefunction values at quadrature points works for gradients
aswell now.

Tested with the implementation of Poisson equation and compared with an
Albert implementation of the same problem. :-) we are close.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@112 419231ca-4411-0410-aba2-fc6cf8c4f609