Added < bool checkInside (Vec .. ) > on class Element. This method returns true if an point is located within the current element.
Furthermore added an interface class BoundaryEntity which provides
information about the boundary. This is a new template parameter for
NeighborIterator. I adapted the interface classes and SGrid and AlbertGrid.

On NeighborIterator the method < bool boundary .. > returns wether we are
at boubdary or not. The method < neighbor > returns true if an neihgbor on this
level exists. The method < boundaryEntity > returns a BoundaryEntity with
information about boundary type and so on, if we are at boundary.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@160 419231ca-4411-0410-aba2-fc6cf8c4f609