Whats new:
global_index:

Added the global_index management for entity<codim=0>. Now each entity<0>
the in AlbertGrid has an global unique index. We need this for adaptive
grids were we only use the leafs. The global_index starts from 0 to
global_size( codim ) .

Each grid should have this global_index, because with that we can
setup an mapping to consecutive indices on each level.
For SGrid we dont have to setup such an mapping because its implicitly
given.

Adaptation:

The Grid class has three adaptation methods, preAdapt() : bool which
returns true if a leaast one entity<0> was marked for coarsening.
adapt() : bool which returns true if a least one entity<0> was refined and
postAdapt() : void.

preAdapt is called before adapt in called and is an interface method to
give the implementer of the grid interface some room for thing special to
one grid. The method adapt adapts the grid( refinement and coarsening ).
postAdapt is for the implementer to clean up some thing after the
adaptation.

On the class Entity<0> we have to methods, mark(int), which marks for
coarsening or refinement and state () : AdaptationState where the user can
get informations about the state on an Entity<0>.

Parallelization:

This is still test status. So far we the LevelIterator is improved for
running over differrent marked entitie<0> ( Intertior, Ghosts ...)
and on each entity we can get indformations about the owner and
PartitionType.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@434 419231ca-4411-0410-aba2-fc6cf8c4f609