Using template specialization I separated the parallel stuff of AMG from the sequential version.
It is now possible to use a purely sequential version of AMG by using
the MatrixAdapter as Operator and SequentialInformation instead of
ParallelInformation. In this case now IndexSets, RemoteIndices and
other stuff needed for communication are created and used. Thus there
is now memory overhead.

TODO: MPI is still needed to compile AMG. For the stable version it
should be possible to compile AMG even if MPI is not present.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@3548 419231ca-4411-0410-aba2-fc6cf8c4f609