In their infinite wisdom, the mighty creators of UG decided that each and every refinement rule be designated by an enumeration value.  And this was so even if two rules were the same except for orientation.
And it was good.

But the forces of evil made that no rule be without exception, and this
is why the UG method MarkForRefinement has a little mystery argument
'side' (known to the elders as 'userdata'), which chooses the orientation
of anisotropic refinement.  But, lo and behold, only if DIM=2 and
the refinement rule is BLUE.

So if poor infidel me wants to have both orientations of BLUE, I have
to add the 'side' parameter to the custom UGGrid mark() method.  Which
hath herewith been done, carrying a default value of 0.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@3756 419231ca-4411-0410-aba2-fc6cf8c4f609