Reimplemented the method geometryInFather() and now it should work properly even when the boundaries are parametrized. This fixes bug 74.
Also there was a bug in operator[] when not in element_mode.  Since
the element vertices are always stored in UG order no matter whether
we directly store coordinates or pointers into the UG data structure,
the usual UG/DUNE vertex renumbering has to be done in non-element-mode
as well.

Last but not least: UGGridEntity contains some big methods.  It is
therefore precompiled into libdune from now on.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@3929 419231ca-4411-0410-aba2-fc6cf8c4f609