Cleaned up the mechanism that lets FieldVectors have size 1 internally when they're instantiated with size 0.
WARNING: this patch contains a tiny semantic change: the public enum
size, which used to contain the internal size of the FieldVector now
contains the size that the vector was instantiated with.  This must
be so because the internal size is an implementation details should
be invisible from the outside.  As long as you didn't work with
size 0 FieldVectors it doesn't matter anyways.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@4304 419231ca-4411-0410-aba2-fc6cf8c4f609