[ConfigParser] Allow to retain older entries when parsing from a file/stream
Often the configuration file is given on the command line.  However,
configuration entries can also be given on the command line, and should
usually take precedence over those from a file.  With these changes it is
possible to read entries from the command line, possibly setting one
configuration entry to the name of a config file in the process.  The
configuration file can then be parsed, setting entries, which have not been
set on the command line.  Entries which already have been set on the command
line will retain their command-line values.

git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@6007 419231ca-4411-0410-aba2-fc6cf8c4f609