Provide the existing parsing facilities for ParameterTrees in a separate file
This code used to be in ConfigParser, which I will deprecate in a second.
The new implementation is inspired by boost: each parser is a global method
in a namespace.  In our case, that namespace is a class ParameterTreeParser,
because it needs to have some private methods (ltrim and rtrim, to be
precise).  On demand, parsers for more formats can be added to this class/file.

I do not insist on any of the details of this new code. However, I wanted
to have the many architectural change before the creation of the 2.2 release
branch.


git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@6346 419231ca-4411-0410-aba2-fc6cf8c4f609