[hash] Make bigunsignedint hashable using Dune::hash
This patch adds the required hooks to enable hashing of bigunsignedint
and makes sure the hasher can be invoked in bigunsignedinttest.

Making bigunsignedint hashable also makes it possible to use hash-based
containers for EntityIDs with YaspGrid and SGrid, as those grids implement
the EntityID as a plain bigunsignedint.

The patch also adds a little test to bigunsignedinttest, trying to hash
a bigunsignedint with Dune::hash as well as any of the detected backends
(std::hash and std::tr1::hash).

Kudos to Steffen Müthing, see FS#1192

git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@7066 419231ca-4411-0410-aba2-fc6cf8c4f609