Fix buggy use of AC_TRY_COMPILE
Thanks to Andreas Lauser for this patch.

Here is his own commit message:

From: Andreas Lauser <and@poware.org>
Date: Mon, 4 Feb 2013 15:34:22 +0100
Subject: [PATCH] fix the test for __attribute__((unused))

it contained two errors which "anihilated" themselfs: the first was a
syntax error in the test itself (the UNUSED macro must be placed
before the identifier of the variable), which caused the compiler to
always return an error. this was compensated by an incorrect
invocation of the AC_TRY_COMPILE macro which featured an empty
argument as third argument instead of the action to be executed if the
test succeeded. this in turn always lead the forth argument (which was
intended by the test to be executed if the test succeeds but was
actually executed if it failed) to be always triggered. In turn this
leads me to the conclusion that DUNE was never compiled using a
compiler which does not support __attribute__((unused)) since this
macro was introduced...



git-svn-id: https://svn.dune-project.org/svn/dune-common/trunk@7105 419231ca-4411-0410-aba2-fc6cf8c4f609